pyproject.toml
[project]
name = "betting-odds-pipeline"
version = "0.1.0"
description = "NBA betting odds data collection and analysis system"
authors = [{name = "Andrew"}]
requires-python = ">=3.11"
dependencies = [
    "sqlmodel>=0.0.14",
    "asyncpg>=0.29.0",
    "aiohttp>=3.9.0",
    "apscheduler>=3.10.4",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "pydantic-settings>=2.1.0",
    "alembic>=1.13.0",
    "tenacity>=8.2.3",
    "structlog>=23.3.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.34.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "pre-commit>=3.5.0",
    "freezegun>=1.5.5",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "seaborn>=0.13.2",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "torch>=2.0.0",
]

[project.scripts]
odds = "cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "storage", "jobs", "cli", "analytics", "alerts"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
